gunyah: Support killable sleep in gh_msgq_recv
When Android begins freezing-of-tasks as part of entering suspend, crosvm,
and thus all VMs managed by crosvm, is frozen. Similarly, userspace
threads which communicate with these VMs via the smmu-proxy kernel driver
will be sent signals, and similarly frozen. But smmu-proxy driver is
currently unable to handle these signals cleanly and return -ERESTARTSYS
to userspace.

Consider the following:
1) smmu-proxy lends a dma-buf to TVM
2) smmu-proxy sends a message to TVM requesting it to map the dma-buf
2a) TVM begins map operation.
3) smmu-proxy calls gh_msgq_recv to wait for a response message from TVM.
4) Android suspend starts
5) smmu-proxy woken from gh_msgq_recv with -ERESTARTSYS

smmu-proxy cannot send a message to TVM to request it to clean up by
virtue of the crosvm thread being frozen. Also, smmu-proxy cannot
'undo' step 1 because it depends on TVM cleanup completing. Furthermore,
when crosvm unfreezes, TVM will attempt to continue the in-progress
operation.

We propose to solve this by:
6) pm_stay_awake() call from smmu-proxy driver
7) gh_msgq_recv sleeps with TASK_KILLABLE | TASK_UNINTERRUPTIBLE

Explanation:
It is not feasible to have the userspace entity managing the smmu-proxy
driver usecase to hold a wakelock because of the refcounted nature of
dma-buf objects. For example, if a dma-buf was shared between N clients,
both userspace and kernel, then all of them would need to hold a wakelock,
so that the owner of the last refcount would prevent suspend when closing
the dmabuf (and thus requesting smmu-proxy driver to send an unmap request
to TVM). It is simpler to hold wakelock in smmu-proxy driver instead.
This wakelock will prevent crosvm from being frozen, or abort suspend if
it is already in progress.

7) will resolve the issue with -ERESTARTSYS mentioned above, if suspend
is already in progress when the wakelock is acquired. Guaranteeing
uninterruption of the existing sequence is also safer from a stability
perspective by not invoking the error paths in N drivers which might
not be expecting ERESTARTSYS to be returned.

wait_event_freezeable was considered, but it does not work because
the comment in __set_task_frozen advises against freezing with
locks held.

Change-Id: I78cd9e2368eb14dc054393b340c81b0e45e5a529
Signed-off-by: Patrick Daly <quic_pdaly@quicinc.com>
Signed-off-by: Piyush Pradeep Jain <quic_piyushpr@quicinc.com>
(cherry picked from commit 932114b195c9272b6cc379f505d4936614c1de55)