usb: dwc3: dwc3-msm-core: Resume SS PHY if in 2/4LN DP mode always
commit 28feebbd6f30 ("usb: dwc3-msm: Allow DP mode only after host mode
has started")

Added a check for dp_state before allowing the SSPHY to be powered down.
However, it avoided updating the logic that was present in determining if
the SSPHY needed to be resumed from the normal resume path.
(dwc3_msm_resume())

This led to a situation where a 2LN DP adapter w/o a SSUSB hub, would
keep the DWC3 wrapper's source clock mux to rely on PIPE clock being
available.  However, when a bus resume scenario occurs, the
dwc3_msm_resume() would avoid powering up the SSPHY, and the xHCI resume
path would be running w/o a source clock to the controller.

Change-Id: I7def6aecd003be386e6daad98463975febc9a4ee
Signed-off-by: Wesley Cheng <quic_wcheng@quicinc.com>