usb: dwc3: dwc3-msm-core: Return 0 from usb_audio_pre_reset()
When usb_reset_device() is called, it executes the pre/post reset
functions of the interface registered. However if the pre_reset returns
non zero, then the interface's need_binding gets set in function
usb_forced_unbind_intf(). This would further result in rebind to be set
to 1 which prevents port_reset from executing.

This is typically seen when ESD usecases on usb headsets. Bus reset
occurs due to ESD and causes audio mute.

Fix this by returning 0 from usb_audio_pre_reset() which would align the
driver to the expectation of the underlying function.

Change-Id: I1bbdaa9ba18de6f5d94da18c854c99a226579abe
Signed-off-by: Udipto Goswami <quic_ugoswami@quicinc.com>