sched/walt: Don't disable LPM bias if pipeline is active
For systems with one prime cpu, under OBET scenario, if pipeline is
active on a prime cpu, then do not disable LPM bias for the prime
cluster.
This prevents the prime cpu to enter deep idle state if it was busy
previously , hence helps in reducing latency for pipeline tasks.
As before, for systems with more than one prime cpus, under OBET
scenario LPM bias will be disabled.

Change-Id: Iab934e7f910a21e54b6dfc4bfda44d9e79720cd5
Signed-off-by: Atul Pant <quic_atulpant@quicinc.com>