xhci: sideband: Adding NULL check while removing endpoint
Currently, the xhci_sideband_remove_endpoint() function agnostic of
weather the sideband is NULL or not. Further the functions extracts the
mutex from the sideband which if is already NULL will result in a kernel
panic.

Fix this by adding a NULL pointer check in
xhci_sideband_remove_endpoint() to make sure if sideband is NULL then
omit further operations and bail out.

Change-Id: Ibb04724403b30d75ec72673d10f4061e1878fb96
Signed-off-by: Udipto Goswami <quic_ugoswami@quicinc.com>