power: supply: Add workaround for CHG_BST bias voltage upon plugin
This workaround for a mysterious PMIC hardware issue avoids a ramdump on
charger connect. In short, if the charger is briefly connected and then
disconnected (~100ms), a transient voltage will appear on CHG_BST and
remain there for the next ~10 seconds. If audio is played during this
time, the process of turning on the Vreg for CHG_BST will fail due to
the pre-existing voltage there. This failure manifests as an RPM abort.

As a workaround, we enable a 10mA pulldown on CHG_BST (internal to the
PMIC) for 50ms, each time Vbus is applied (the charger is plugged in).

This workaround is enabled in the DT via "qcom,chg-bst-pulldown-wa".
The pulldown enable duration defaults to 50ms and can be overridden via
"qcom,chg-bst-pulldown-wa-duration-ms".

Change-Id: I569d9971c448436f13896c25c40f3704cf7d3e4e
Signed-off-by: Ayyagari Ushasreevalli <quic_aushasre@quicinc.com>