gunyah: Cancel vcpu wq only when vCPUs are running in workqueue mode
Cancelling a work for which INIT_WORK() was not called is incorrect.
This can arise in two scenarios. First, when VM exits before the vCPU
FDs are dropped and vCPUs don't get chance enter in workqueue mode.
Second, for the VMs which do not have the keep_running property set,
vCPUs never enter workqueue mode.
This patch adds check to cancel vcpu work only if the vCPUs are running
in workqueue mode. Also fixes a double queue_work().

Change-Id: Id21038de522f2eace0f54d2fcc6dc748f4c8d4ef
Signed-off-by: Hrishabh Rajput <hrishabh.rajput@oss.qualcomm.com>