soc: qcom: minidump: Correct CPU ID in task info
The task_struct->on_cpu field is used to indicate whether a task is
currently running on a CPU, but it does not specify which CPU the
task is running on. To correct this, we should use thread_info->cpu,
which accurately identifies the CPU on which the task is located.

Change-Id: If334933921353d8e2ee4b41389e1af86e27f41cd
Signed-off-by: Huang Yiwei <quic_hyiwei@quicinc.com>
Signed-off-by: Souradeep Chowdhury <quic_schowdhu@quicinc.com>