modules.list.msm.qcs610: Add spi module to first stage
The camera lens driver depends on the SPI driver for proper
initialization.This timing mismatch causes a binding failure
in the lens driver due to the unavailability
of the SPI interface.

With this change the SPI driver is available before the lens
driver attempts to bind, thereby resolving the initialization
issue.

Change-Id: I489339c52cbfa88e9326b719d6c7a49a907d457a
Signed-off-by: Nandi Bhoopathi <quic_nandibho@quicinc.com>